Remove duplicate Cart Razor Page
